Don't drain completion hints for unrealated requests
This is important in cases a task uses submit_unordered, iterates over
the stream and submits another request in the loop body: In this case,
two child futures (completion of requested items and request in loop
body) are active. If one of those consumes the completion hints for the
other one it will not make progress.