Batch requests in volume::mean
It seems like for large volumes there is a (somewhat) large scheduling
overhead (handling of completed tasks/all requested data in
`submit_unordered_with_data`. For now we thus just batch in the only
operations that collect a very large number of bricks. In the future we
may want to find a more general solution.