Parallelize volume::rescale
For this `ThreadReadHandle` and `ThreadWriteHandle are introduced in
addition to the existing `ReadHandle` and `WriteHandle` which can be
passed to threads and then returned to the main thread for completion
(marking initilized/dropping to decrement read handle counter).
Ultimately this allows high throughput and low latency implementation of
operators using the `then` method of async streams.