Lock entries upon request
This ensures that the result of a computation cannot be freed before the
requestor get a change to read it. Previously, this was a problem with
grouped requests where one item of the group was fulfilled earlier that
others and then freed to make room for the next ones or its
dependencies, resulting in group requests never being completed.