Allow single-item granularity operators
In some cases (e.g. renderers that dynamically request bricks from
shader code) the complexity of handling multiple item requests in
parallel is not worth it. One would have to reimplement a smart
scheduler within the operator. It is thus likely the best way forward to
let the global scheduler do this.