Change sliceviewer to request bricks from shader
Instead of a fixed region, a brick index buffer is built from the
requested bricks. One immediate benefit from this is that we no longer
request the full AABB of the slice in voxel space which is beneficial
for non-axis aligned slices.

There is some work to be done, however: The index is created and updated
per task (i.e., per tile of the output image) which is duplicate work.
Also, bricks are never released from memory until the tile is fully
rendered.