Allow specification of a maximum gpu storage size
Ideally the allocator would just reliably return an oom error when the
heap memory is exhausted, but this does not work in practice:
 - On integrated graphics cards (at least on linux with intel igpu) the
   video memory resides in main memory, which results in the application
   to just be oom-killed when it allocates to much.
 - It can be the case (probably again depending the driver, at least for
   the amd mesa driver) that we consome so much vram that there is no
   space left for, for example, command buffer submission. I highly
   doubt that we can reliably detect this and release memory in all
   cases. If we can, this would be very tedious.

There is also VK_EXT_memory_budget, but it does not appear to be
supported everywhere. Also, the gpu_allocator crate does not expose the
memory heaps currently, so we cannot really make use of this
information.

The solution for now: In case the driver does not play well with the
current on-oom-mem-release scheme, the user can just specify a global
maximum gpu storage size. Currently, for simplicity, this applies to all
devices, but this can be changed in the future.