Lazily free bricks referenced in indices
For now this happens in a two-step process: If the free budget could not
be reached, bricks referenced by indices are unreferenced. This
(potentially) make them available for deletion in the next garbage
colleciton phase.

This approach is not perfect (e.g. may the gargabe collection fail
theoretically if all bricks are referenced in indices and thus nothing
can be freed), but this is highly unlikely. Garbage collection will
change some time soon anyway, though, so I think this is fine for now.