Change raycaster request structure once again
As part of this, we revert BrickRequestTable to be a temporary ressource
again. This has the advantage that we don't lose request performance
during small camera movements (which would throw away the cached request
table).

Instead, now we first render, then check if we are done. If not, bricks
are requested and only then we check for timeouts.