Get rid of Operator lifetime
As a consequence of this, all Operators (or rather the compute
functions) and all state must alive for the 'static lifetime. In
essence, this means that we pass state via Rc or similar into the
operators. While this is somewhat a step back in terms of the rust API
(because rc'ed objects may in theory escape and leave the state behind
and since we need more interior mutability) this is sadly unavoidable in
preparation for a python API.