Use mostly stack data instead of ScalarOperator
For almost all use cases there is no benefit in using a ScalarOperator,
but multiple downsides (e.g., we cannot determine the size of the LOD
stack dynamically).

If we generate a ScalarOperator (at the moment the only possibility is
via the mean operator), we can always just evaluate this expression
before constructing and evaluating the rest of the network. Evaluation
of this scalar mean will be a bottleneck anyways.