Avoid spawning the same data task twice
This could previously happen because we were erroneously checking if the
exact Request had been initiated before, but this should depend on the
more general data id. (This is because we don't know in what format a
task may generate data and also not with what visibility (for GPU
data)).