Fix multiple scheduling issues
It's hard to split this into logically separate commits because the
previous version "kind-of-almost" worked, but any intermediate changes
on their own made the schedlung far more unstable (i.e. prone to
deadlocks and creating illegal states). The current state is still not
great (task_graph.rs and the scheduling in runtime.rs is far too complex
for my liking), but appears to work for now... Maybe some day I'll get
an idea for a more formal way of modelling the dependency system.