Fix reseting of color state in raycaster
We now use the (potentially cached) input frame as a separate buffer and
use the color cache for compositing. Only when we composite a value != 0
we write it out. This has the added benefit that during refinement there
is a smaller chance that the rendered preview value is overwritten by a
black pixel (e.g. because the volume is empty at the border).