Calculate proper mean statistics at the border for rw weight functions
Previously, we used mean filter with repeat strategy which resulted in a
bias for the values at the very border of the volume. Now we pad with
zeros (or equivalently: only sum up to the border) and normalize with
the actual number of values inside the filter (not counting those values
hit by the filter outside of the volume).