Always allow capacity increase for transfer buffers
They are allocated on a different heap than the vast majority of other
allocations (GpuOnly). Our model of a single capacity for the whole gpu
doesn't quite fit the bill (due to different heaps for different memory
sizes), so a redesign might be in order in general, but this helps the
situation for now.